aks_seldon/model/requirements.txt
azure-identity
azure-keyvault-secrets
azure-core
azure-ai-textanalytics
seldon-core